+++ src/math/matrix33.cc
@@ -61,14 +61,14 @@ Matrix33 Matrix33::rotationAroundZ(const real angle)
 }
 
 
-Matrix33 Matrix33::rotationAroundPrincipalAxis(unsigned i, const real angle)
+Matrix33 Matrix33::rotationAroundPrincipalAxis(size_t i, const real angle)
 {
     real c = cos(angle);
     real s = sin(angle);
     
     i %= 3;
-    int j = (i+1)%3;
-    int k = (i+2)%3;
+    size_t j = (i+1)%3;
+    size_t k = (i+2)%3;
     
     Matrix33 res(0, 1);
     res(j,j) = c;
